Added main comICA functions
Added concatenate_epochs()  and distance_matrix() functions. The first function is used to create a spatial concatenation of the epochs from the different recordings. The concatenation is done with N=2 recordings only, but can be done iteratively to include the N>2 case. The distance matrix function can be set to calculate either the whitened distance matrix  or the unwhitened one. It contains the calculation of the covariance matrix used for whitening.